move files to core folder and fix all imports
