+++ src/core/App.js
