+++ src/core/functions.js
