+++ src/core/routes.js
@@ -1,8 +1,8 @@
 import { createHashRouter, createRoutesFromElements, Navigate, Route } from "react-router-dom";
-import MovieDetails from "./features/movies/MovieDetails";
-import MoviesList from "./features/movies/MoviesList";
-import PeopleList from "./features/people/PeopleList";
-import PersonDetails from "./features/people/PersonDetails";
+import MovieDetails from "../features/movies/MovieDetails";
+import MoviesList from "../features/movies/MoviesList";
+import PeopleList from "../features/people/PeopleList";
+import PersonDetails from "../features/people/PersonDetails";
 
 export const toPopularMovies = "/movies";
 export const toMovieDetails = "/movies/movie/:id";
